+++ pkg/resolver/zoner.go
@@ -235,11 +235,16 @@ func (d *Zoner) getTagsForHostedZone(cluster Cluster) map[string]string {
 }
 
 func (d *Zoner) getWorkloadClusterDnsRecords(workloadClusterHostedZoneName string, cluster Cluster) []DNSRecord {
+	wildcardCNAMETarget := cluster.WildcardCNAMETarget
+	if wildcardCNAMETarget == "" {
+		wildcardCNAMETarget = fmt.Sprintf("ingress.%s", workloadClusterHostedZoneName)
+	}
+
 	dnsRecords := []DNSRecord{
 		{
 			Kind:   DnsRecordTypeCname,
 			Name:   fmt.Sprintf("*.%s", workloadClusterHostedZoneName),
-			Values: []string{fmt.Sprintf("ingress.%s", workloadClusterHostedZoneName)},
+			Values: []string{wildcardCNAMETarget},
 		},
 	}